src/main/java/pl/tkaczyk/projektstudia/controller/TodoController.java
package pl.tkaczyk.projektstudia.controller;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import pl.tkaczyk.projektstudia.model.Todo;
import pl.tkaczyk.projektstudia.repository.TodoRepository;

import java.util.List;

@RestController
public class TodoController {
    private TodoRepository todoRepository;

    public TodoController(TodoRepository todoRepository) {
        this.todoRepository = todoRepository;
    }

    @GetMapping("/users/{username}/todos")
    public List<Todo> retrieveTodos(@PathVariable String username){
        return todoRepository.findTodoByUsername(username);
    }
    @GetMapping("/users/{username}/todos/{id}")
    public Todo retreieveTodo(@PathVariable int id, @PathVariable String username) {
        return todoRepository.findById(id).get();
    }

    @DeleteMapping("/users/{username}/todos/{id}")
    public ResponseEntity<Void> deleteTodo(@PathVariable int id, @PathVariable String username) {
        todoRepository.deleteById(id);
        return ResponseEntity.noContent().build();
    }

    @PutMapping("/users/{username}/todos/{id}")
    public Todo updateTodo(@PathVariable String username, @PathVariable int id, @RequestBody Todo todo) {
        todoRepository.save(todo);
        return todo;
    }

    @PostMapping("/users/{username}/todos")
    public Todo createTodo(@PathVariable String username, @RequestBody Todo todo) {
        todo.setUsername(username);
        return todoRepository.save(todo);
    }
}
